requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crypto==1.4.1
cryptography==42.0.5
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.110.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
mongoengine==0.28.2
Naked==0.1.32
packaging==24.0
pluggy==1.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
pymongo==4.6.3
pytest==8.1.1
pytest-mock==3.14.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
shellescape==3.8.1
sniffio==1.3.1
starlette==0.37.2
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
